AI-Service/requirements.txt
# FastAPI & Web
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0
python-dotenv==1.0.0
python-multipart==0.0.6

# LangChain - Core
langchain==0.1.0
langchain-core==0.1.10
langchain-google-genai==0.0.6
langchain-openai==0.0.3
langchain-community==0.0.13
langgraph==0.0.20

# Google AI
google-generativeai==0.3.2

# Computer Vision
mediapipe==0.10.9
opencv-python==4.9.0.80
pillow==10.2.0
numpy>=1.24.0

# Real-time
python-socketio==5.11.0
websockets==12.0

# Storage & Database
redis==5.0.1
motor==3.3.2
pymongo==4.6.1

# HTTP Client
httpx==0.26.0
aiohttp==3.9.1

# Utilities
tenacity==8.2.3
